made fixes requested in the pr
