+++ practice-app/app/templates/index.html
@@ -41,7 +41,7 @@
                     <img src="static/images/pokemon.jpg" alt=""/>
                 </div>
                 <div class="content">
-                    <h3><a href="{{ url_for('PokemonPage') }}"> Pokemon </a></h3>
+                    <h3><a href="{{ url_for('pokemon_page') }}"> Pokemon - Omer Bahadiroglu </a></h3>
                     <p>Can you find an unclaimed pokemon? Try it. Find it first and own it.</p>
                 </div>
             </div>
